requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.8
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==65.0.2
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
oauthlib==3.2.2
openai==1.35.10
packaging==24.1
persian-tools==0.0.11
platformdirs==4.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.3
